requirements.txt
pyocr
pillow
google-api-python-client
livestreamer
nose
requests
